+++ src/main.go
@@ -1,15 +1,25 @@
 package main
 
 import (
+	"bufio"
+	"bytes"
+	"encoding/json"
 	"fmt"
+	"io/fs"
 	"log"
+	"net/http"
 	"os"
 	"os/exec"
 	"path"
+	"path/filepath"
 	"reflect"
+	"sort"
+	"strconv"
 	"strings"
+	"time"
 
 	"github.com/bestmethod/inslice"
+	"github.com/google/uuid"
 	flags "github.com/rglonek/jeddevdk-goflags"
 )
 
@@ -186,7 +196,180 @@ func earlyProcessV2(tail []string, initBackend bool) (early bool) {
 	if b != nil {
 		b.WorkOnServers()
 	}
+	go telemetry()
+	/*
+		err = telemetry()
+		if err != nil {
+			log.Printf("TELEMETRY:%s", err)
+		} else {
+			log.Print("TELEMETRY:OK")
+		}
+	*/
 	return false
+}
+
+func telemetry() error {
+	if len(os.Args) < 2 {
+		return nil
+	}
+
+	// only enable if a feature file is present and belongs to Aerospike internal users
+	if a.opts.Cluster.Create.FeaturesFilePath == "" {
+		return nil
+	}
+	enableTelemetry := false
+	err := filepath.WalkDir(string(a.opts.Cluster.Create.FeaturesFilePath), func(path string, d fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+		if d.IsDir() {
+			return nil
+		}
+		if !d.Type().IsRegular() {
+			return nil
+		}
+		file, err := os.Open(path)
+		if err != nil {
+			return err
+		}
+		defer file.Close()
+		scanner := bufio.NewScanner(file)
+		for scanner.Scan() {
+			if err = scanner.Err(); err != nil {
+				return err
+			}
+			line := strings.ToLower(strings.Trim(scanner.Text(), "\r\n\t "))
+			if strings.HasPrefix(line, "account-name") && (strings.HasSuffix(line, "aerospike") || strings.Contains(line, " aerospike") || strings.Contains(line, "aerospike_test") || strings.Contains(line, "\taerospike")) {
+				enableTelemetry = true
+			}
+		}
+		if err := scanner.Err(); err != nil {
+			return err
+		}
+		return nil
+	})
+	if err != nil {
+		return err
+	}
+	if !enableTelemetry {
+		return nil
+	}
+
+	// resolve telemetry directory
+	home, err := os.UserHomeDir()
+	if err != nil {
+		return err
+	}
+	telemetryDir := path.Join(home, ".aerolab.telemetry")
+
+	// check if telemetry is disabled
+	if _, err := os.Stat(path.Join(home, ".aerolab.disabletelemetry")); err == nil {
+		return err
+	}
+
+	// create telemetry dir
+	if _, err := os.Stat(telemetryDir); err != nil {
+		err = os.MkdirAll(telemetryDir, 0755)
+		if err != nil {
+			return err
+		}
+	}
+
+	// generate uid for telemetry if it doesn't exist
+	var uuidx []byte
+	uuidFile := path.Join(telemetryDir, "uuid")
+	if _, err := os.Stat(uuidFile); err != nil {
+		uuidx = []byte(uuid.New().String())
+		err = os.WriteFile(uuidFile, uuidx, 0644)
+		if err != nil {
+			return err
+		}
+	} else {
+		uuidx, err = os.ReadFile(uuidFile)
+		if err != nil {
+			return err
+		}
+	}
+
+	// get telemetry items
+	telemetryFiles := []string{}
+	err = filepath.WalkDir(telemetryDir, func(path string, d fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+		if d.IsDir() {
+			return nil
+		}
+		if !d.Type().IsRegular() {
+			return nil
+		}
+		if !strings.HasPrefix(d.Name(), "item-") {
+			return nil
+		}
+		telemetryFiles = append(telemetryFiles, path)
+		return nil
+	})
+	if err != nil {
+		return err
+	}
+
+	// if telemetry count > 1000, abort
+	if len(telemetryFiles) > 1000 {
+		return nil
+	}
+
+	// sort telemetryFiles
+	sort.Strings(telemetryFiles)
+
+	// add current command to telemetry
+	item := telemetryItem{
+		UUID:    string(uuidx),
+		Time:    time.Now().UnixMicro(),
+		CmdLine: os.Args[1:],
+	}
+	telemetryString, err := json.Marshal(item)
+	if err != nil {
+		return err
+	}
+	newFile := path.Join(telemetryDir, "item-"+strconv.Itoa(int(item.Time)))
+	err = os.WriteFile(newFile, telemetryString, 0644)
+	if err != nil {
+		return err
+	}
+	telemetryFiles = append(telemetryFiles, newFile)
+
+	// ship telemetryFiles oldest to newest and remove from disk
+	for _, file := range telemetryFiles {
+		if err = telemetryShip(file); err == nil {
+			err = os.Remove(file)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+func telemetryShip(file string) error {
+	contents, err := os.ReadFile(file)
+	if err != nil {
+		return err
+	}
+	url := "https://us-central1-aerospike-gaia.cloudfunctions.net/aerolab-telemetrics"
+	ret, err := http.Post(url, "application/json", bytes.NewReader(contents))
+	if err != nil {
+		return err
+	}
+	if ret.StatusCode < 200 || ret.StatusCode > 299 {
+		return fmt.Errorf("returned ret code: %d:%s", ret.StatusCode, ret.Status)
+	}
+	return nil
+}
+
+type telemetryItem struct {
+	CmdLine []string
+	UUID    string
+	Time    int64
 }
 
 func writeConfigFile() error {